requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
altair==5.1.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
dateparser==1.2.0
fonttools==4.45.0
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.1
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydeck==0.8.1b0
Pygments==2.17.1
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-binance==0.7.5
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
service-identity==23.1.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
streamlit==1.28.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
websockets==12.0
yarl==1.9.3
zipp==3.17.0
zope.interface==6.1
